moved job builder helpers to helpers.go added helper for parsing config files
