+++ core/pkg/launch/job_builder_test.go
@@ -1,15 +1,13 @@
-package launch_test
+package launch
 
 import (
 	"os"
 	"path/filepath"
-	"reflect"
 	"testing"
 
 	"github.com/segmentio/encoding/json"
 
 	"github.com/stretchr/testify/assert"
-	. "github.com/wandb/wandb/core/pkg/launch"
 	"github.com/wandb/wandb/core/pkg/observability"
 	"github.com/wandb/wandb/core/pkg/service"
 	"google.golang.org/protobuf/types/known/wrapperspb"
@@ -889,72 +887,4 @@ func TestUtilFunctions(t *testing.T) {
 		assert.Nil(t, err)
 		assert.Equal(t, "a/notebook.ipynb", path)
 	})
-}
-
-func TestFilterDataStructure(t *testing.T) {
-	tests := []struct {
-		name       string
-		ds         map[string]interface{}
-		keys       []string
-		filterIn   bool
-		wantResult map[string]interface{}
-	}{
-		{
-			name: "FilterOutSingleKey",
-			ds: map[string]interface{}{
-				"config": "value1",
-				"data":   "value2",
-			},
-			keys:       []string{"config"},
-			filterIn:   false,
-			wantResult: map[string]interface{}{"data": "value2"},
-		},
-		{
-			name: "FilterOutNestedKey",
-			ds: map[string]interface{}{
-				"config": map[string]interface{}{
-					"key": "value",
-				},
-				"data": "value2",
-			},
-			keys:       []string{"config.key"},
-			filterIn:   false,
-			wantResult: map[string]interface{}{"config": map[string]interface{}{}, "data": "value2"},
-		},
-		{
-			name: "FilterInSingleKey",
-			ds: map[string]interface{}{
-				"config": "value1",
-				"data":   "value2",
-			},
-			keys:       []string{"config"},
-			filterIn:   true,
-			wantResult: map[string]interface{}{"config": "value1"},
-		},
-		{
-			name: "FilterInNestedKey",
-			ds: map[string]interface{}{
-				"config": map[string]interface{}{
-					"key": "value",
-				},
-				"data": "value2",
-			},
-			keys:     []string{"config.key"},
-			filterIn: true,
-			wantResult: map[string]interface{}{
-				"config": map[string]interface{}{
-					"key": "value",
-				},
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			gotResult := FilterDataStructure(tt.ds, tt.keys, tt.filterIn)
-			if !reflect.DeepEqual(gotResult, tt.wantResult) {
-				t.Errorf("filterDataStructure() = %v, want %v", gotResult, tt.wantResult)
-			}
-		})
-	}
 }
